Reduce compile branches in Series
All the data types are backed by a physical type in memory e.g. Date32 -> i32, Date64 -> i64.

Series lead to code implementations of all traits. Whereas there are a lot of duplicates due to
data types being backed by the same physical type. In this commit we reduce compile times by
opting for a little more run time cost. We cast to the physical type -> apply the operation and
(depending on the result) cast back to the original type