Improve compatibility in supported options between pyarrow and polars CSV reading code.
Improve compatibility in supported options between pyarrow and
polars CSV reading code.

The following options are now also supported when reading data with
the pyarrow code:
  - "projection"
  - "columns"
  - "ignore_errors"
  - "new_columns"

When reading CSV data with pyarrow and when column names need to
be auto generated, they are automatically converted from:
  'f0', 'f1', ... columns names autogenated by pyarrow
to:
 'column_1', 'column_2', ... for polars.
So that the auto generated column names are not changing when
adding an extra option to read_csv that is not supported (yet)
by the pyarrow CSV reading code.