Proposal: Remove date_like_to_physical
This function was only used in arithmetic methods in pl.Series. Not sure what we would expect, what should the answer be when we subtract two dates, or a date and a time, given we do not have a timedelta data type. No tests are hitting this logic either. So I propose to remove it, but I may be missing something here..