Preserve types of DataFrame subclasses (#2859)
This PR introduces the necessary changes such that all methods of DataFrame and LazyFrame which return a new DataFrame or LazyFrame objects, respectively, preserve the type when these classes have been inherited from.

This PR implements solutions for tasks 1-3 as outlined in #2846. Preservation of data types after roundtrips from DataFrame to GroupBy/LazyFrame and back still remains to be done as well. I have a solution in mind for task 4 and 5 as outlined in the issue, but I will introduce that one in a separate PR in order to make the review process a bit simpler relaxed