Replace OOB slice indexing with spare_capacity_mut (#3737)
Previously, this code attempted to access the uninitialized region in a
Vec between len and capacity by doing unchecked indexing into a
zero-length slice. It is always UB to do this, and this can be detected
by running the tests with -Zbuild-std to turn on the standard library's
debug assertions, which add bounds checking to get_unchecked.
Vec::spare_capacity_mut is the right API for doing this sort of
operation.